% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonHelper.R
\name{comparison_cellBased}
\alias{comparison_cellBased}
\title{Compare two cartridge case scans using the cell-based comparison procedure}
\usage{
comparison_cellBased(
  reference,
  target,
  thetas = seq(-30, 30, by = 3),
  numCells = c(8, 8),
  returnX3Ps = TRUE,
  sideLengthMultiplier = 3
)
}
\description{
Compare two cartridge case scans using the cell-based comparison procedure
}
