% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonHelper.R
\name{comparison_fullScan}
\alias{comparison_fullScan}
\title{Compare two cartridge case scans by estimating the alignment of the full
scans}
\usage{
comparison_fullScan(reference, target, thetas = seq(-30, 30, by = 3))
}
\description{
Compare two cartridge case scans by estimating the alignment of the full
scans
}
