% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscanCellFeatures.R
\name{densityBasedClusters}
\alias{densityBasedClusters}
\title{Calculate clusters using density-based clustering methods}
\usage{
densityBasedClusters(x, y, eps, minPts, method = "dbscan")
}
\description{
Calculate clusters using density-based clustering methods
}
