% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBasedCellFeatures.R
\name{estimatedRotation}
\alias{estimatedRotation}
\title{Estimate aligning rotation based on high-density translation values}
\usage{
estimatedRotation(x, y, theta)
}
\description{
Estimate aligning rotation based on high-density translation values
}
