% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCalculationHelpers.R
\name{feature_aLaCarte}
\alias{feature_aLaCarte}
\title{Calculate similarity features based on the output of the comparison procedure}
\usage{
feature_aLaCarte(
  comparisonData,
  features = "all",
  id_cols = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{comparisonData}{tibble such as one returned by the
\code{comparison_cellBased()} or \code{comparison_fullScan()} functions that contains
results from the cell-based or full scan comparison procedure}

\item{features}{character vector containing the features to be calculated.}

\item{id_cols}{variable(s) to group by prior to calculating the summary
statistics}

\item{quiet}{Boolean to suppress function messages}

\item{...}{additional parameters for the feature calculation functions}
}
\description{
Calculate similarity features based on the output of the comparison procedure
}
