% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscanCellFeatures.R
\name{feature_densityBasedClusterSize}
\alias{feature_densityBasedClusterSize}
\title{Compute the difference in density-estimated translations between two
comparison directions}
\usage{
feature_densityBasedClusterSize(cluster, direction, imputeVal = 0)
}
\description{
Compute the difference in density-estimated translations between two
comparison directions
}
