% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBasedCellFeatures.R
\name{feature_densityBasedThetaDiff}
\alias{feature_densityBasedThetaDiff}
\title{Compute the difference in density-estimated rotations between two comparison
directions}
\usage{
feature_densityBasedThetaDiff(estimatedThetas, direction, imputeVal = 180)
}
\description{
Compute the difference in density-estimated rotations between two comparison
directions
}
