% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBasedCellFeatures.R
\name{feature_densityBasedTranslationDiff}
\alias{feature_densityBasedTranslationDiff}
\title{Compute the difference in density-estimated translations between two
comparison directions}
\usage{
feature_densityBasedTranslationDiff(x, y, cluster, direction, imputeVal = 1000)
}
\description{
Compute the difference in density-estimated translations between two
comparison directions
}
