% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationFeatures.R
\name{feature_registration_all}
\alias{feature_registration_all}
\title{Calculate all seven registration-based features}
\usage{
feature_registration_all(comparisonData, id_cols = NULL)
}
\arguments{
\item{comparisonData}{a data frame resulting from a comparison of two
cartridge cases (for example, output of the \code{comparison_cellBased} or
\code{comparison_fullScan} functions)}

\item{id_cols}{variable(s) to group by prior to calculating the summary
statistics}
}
\value{
A data frame containing the seven registration-based features: 1.
Mean CCF value 2. Standard deviation (S.D.) of the CCF values 3. Mean
pairwise-complete correlation 4. S.D. of the pairwise-complete correlation
values 5. S.D. of the horizontal translation values 6. S.D. of the vertical
translation values 7. S.D. of the rotation values
}
\description{
Calculate all seven registration-based features
}
\note{
This function can be used on comparison data from a full-scan or
cell-based comparison. For a full-scan comparison, we recommend using only
the average CCF and average pairwise-complete correlation
}
