% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationFeatures.R
\name{feature_registration_summary}
\alias{feature_registration_summary}
\title{Calculate a summary statistic of a variable at the estimated registration}
\usage{
feature_registration_summary(
  cellIndex,
  direction,
  fft_ccf,
  summaryVar,
  summaryFun = mean,
  imputeVal = -1
)
}
\arguments{
\item{cellIndex}{tibble column containing cell indices}

\item{direction}{tibble column indicating whether the associated row came
from the "reference vs. target" or "target vs. reference" comparison}

\item{fft_ccf}{tibble column containing cross-correlation function values}

\item{summaryVar}{tibble column that is to be summarized}

\item{summaryFun}{function that will be used to summarize the values in the
\code{summaryVar} column}

\item{imputeVal}{value to return in case the summary function results in a
non-numeric (e.g., NA, NULL) value}
}
\value{
A numeric value resulting from passing to \code{summaryFun} values in the
\code{summaryVar} column (at which the CCF is maximized per cell index &
direction)
}
\description{
Calculate a summary statistic of a variable at the estimated registration
}
