% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationFeatures.R
\name{feature_registration_all}
\alias{feature_registration_all}
\alias{feature_registration_summary}
\title{Calculate features based on the cell-based or full scan registration
procedures}
\usage{
feature_registration_all(comparisonData, id_cols = NULL)

feature_registration_summary(
  cellIndex,
  direction,
  fft_ccf,
  summaryVar,
  summaryFun = mean,
  imputeVal = -1
)
}
\arguments{
\item{comparisonData}{tibble such as one returned by the
\code{comparison_cellBased()} or \code{comparison_fullScan()} functions that contains
results from the cell-based or full scan comparison procedure}

\item{id_cols}{variable(s) to group by prior to calculating the summary
statistics}

\item{cellIndex}{tibble column containing cell indices}

\item{direction}{tibble column indicating whether the associated row came from
the "reference vs. target" or "target vs. reference" comparison}

\item{fft_ccf}{tibble column containing cross-correlation function values}

\item{summaryVar}{tibble column that is to be summarized}

\item{summaryFun}{function that will be used to summarize the values in the
\code{summaryVar} column}

\item{imputeVal}{value to return if the feature calculation results in a
non-numeric (i.e., NA, NULL) value}
}
\description{
The \code{feature_registration_all} function calculates the seven
registration-based features in one call. The \code{feature_registration_summary}
function is an exported helper.
}
\note{
The \code{feature_registration_all} function can be used on comparison data
from a full-scan or cell-based comparison. For a full-scan comparison, we
recommend using only the average CCF and average pairwise-complete
correlation.
}
\examples{
data("K013sA1","K013sA2")

compData_cellBased <- comparison_cellBased(reference = K013sA1,
                                           target = K013sA2,
                                           thetas = c(-3,0,3))

feature_registration_all(compData_cellBased)

compData_fullScan <- comparison_fullScan(reference = K013sA1,
                                         target = K013sA2,
                                         thetas = c(-3,0,3))

feature_registration_all(compData_fullScan) \%>\%
  dplyr::select(ccfMean,pairwiseCompCorMean)

}
