% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBasedCellFeatures.R
\name{feature_densityBasedClusterSize}
\alias{feature_densityBasedClusterSize}
\title{Compute the average density-based cluster sizes between two comparison
directions}
\usage{
feature_densityBasedClusterSize(cluster, direction, imputeVal = 0)
}
\description{
Compute the average density-based cluster sizes between two comparison
directions
}
