% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualDiagnosticFeatures.R
\name{feature_visualDiagnostic_all}
\alias{feature_visualDiagnostic_all}
\alias{feature_visualDiagnostic_neighborhoodSizeSummary}
\alias{feature_visualDiagnostic_scanDifferenceCor}
\title{Calculate features based on visual diagnostic plots}
\usage{
feature_visualDiagnostic_all(
  comparisonData,
  thresholdMultiplier = 1,
  id_cols = NULL
)

feature_visualDiagnostic_neighborhoodSizeSummary(
  cellHeightValues,
  alignedTargetCell,
  summaryFun = mean,
  thresholdMultiplier = 1
)

feature_visualDiagnostic_scanDifferenceCor(
  cellHeightValues,
  alignedTargetCell,
  thresholdMultiplier = 1
)
}
\arguments{
\item{comparisonData}{tibble such as one returned by the
\code{comparison_cellBased()} or \code{comparison_fullScan()} functions that contains
results from the cell-based or full scan comparison procedure}

\item{thresholdMultiplier}{the default filtering threshold is equal to the
standard deviation of the joint x3p1 and x3p2 surface matrix values. This
argument can be used to scale this threshold}

\item{id_cols}{column names in the comparisonData tibble that uniquely
identify each observation. These are returned along with the computed
features}

\item{cellHeightValues}{list/tibble column of x3p objects containing a
reference scan's cells (as returned by \code{comparison_cellBased()} or
\code{comparison_fullScan()})}

\item{alignedTargetCell}{list/tibble column of x3p objects containing a target
scan's aligned cells (as returned by \code{comparison_cellBased()} or
\code{comparison_fullScan()})}

\item{summaryFun}{function that will be used to summarize the neighborhood
sizes}
}
\description{
The \code{feature_visualDiagnostic_all()} calculates the six visual
diagnostic-based features in one call. Calculate each feature individually
with \code{feature_visualDiagnostic_neighborhoodSizeSummary()} and
\code{feature_visualDiagnostic_scanDifferenceCor()}.
}
\note{
The \code{feature_visualDiagnostic_all} function can be used on comparison
data from a full-scan or cell-based comparison. For a full-scan comparison,
note that the standard deviation features will always be NA.
}
\examples{
data("K013sA1","K013sA2")

compData_cellBased <- comparison_cellBased(reference = K013sA1,
                                           target = K013sA2,
                                           thetas = c(-3,0,3))

compData_cellBased \%>\%
  dplyr::group_by(cellIndex,direction) \%>\%
  feature_visualDiagnostic_all()

compData_fullScan <- comparison_fullScan(reference = K013sA1,
                                         target = K013sA2,
                                         thetas = c(-3,0,3))

compData_fullScan \%>\%
  dplyr::group_by(direction) \%>\%
  feature_visualDiagnostic_all() \%>\%
  dplyr::select(-c(neighborhoodSizeAve_sd,
            neighborhoodSizeSD_sd,
            differenceCor_sd))

}
